pyproject.toml
[tool.poetry]
name = "ai-world"
version = "0.1.0"
description = "AI Agents Framework with FastAPI and LangChain"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
langchain = "0.0.325"
langchain-community = "0.0.10"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sse-starlette = "^1.8.2"
openai = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
